Move email sending code into models
